fixed mantis #0015440: Fragenpools lassen sich nicht dem Test hinzufügen
git-svn-id: http://svn.ilias.de/svn/ilias/trunk@57989 21b2c9ec-7c21-0410-8b45-9bfb8ed2bfc5

+++ Modules/Test/classes/class.ilTestRandomQuestionSetStagingPoolBuilder.php
@@ -67,7 +67,17 @@ class ilTestRandomQuestionSetStagingPoolBuilder
 		while( $row = $this->db->fetchAssoc($res) )
 		{
 			$question = assQuestion::_instanciateQuestion($row['qst_fi']);
-			$question->delete($row['qst_fi']);
+
+			if( $question instanceof assQuestion )
+			{
+				$question->delete($row['qst_fi']);
+			}
+			else
+			{
+				$GLOBALS['ilLog']->writeWarning(
+					"could not delete staged random question (ref={$this->testOBJ->getRefId()} / qst={$row['qst_fi']})"
+				);
+			}
 		}
 
 		$query = "DELETE FROM tst_rnd_cpy WHERE tst_fi = %s";
